Fix the bug that keywords and variables share the same name
A new state is added into the internal parser of REX compiler. When it's called to parse the preserved string from ompparser, the internal lexer enters the new state specifically for expressions. After parsing, the state is popped out.

When all the OpenMP constructs can be handled by the new AST constructor, we can remove all the rules not related to the expression parsing. For now, they should be kept.

Issue https://github.com/passlab/rexompiler/issues/3 is fixed. A new branch may be created later based on this commit, in which the internal grammar files only have the expression parsing module.