Add new function to parse the expressions in IF clause
For this kind of clauses, the dependency on OmpAttribute has been removed.
The reason why we have to pass the directive object into expression parser is that during SgExpression creation, the parser has to check the symbol table to make sure any variable shown in the pragma has been declared properly.

The ownership of expression in the IF clause is set up in the AST contructor now. It's assigned to the corresponding clause instead of the directive. After the changes, it can still pass the test case parallel-if.c with minor modification.