Update the insertion workflow of kmpc global id
The global id won't be available anytime before outlining function. We need to use a different condition to determine whether the variable is available or not.

We need to add the variables "__global_id" and "__bound_id to the allowlist of variable checking in AST outlining function.
SINGLE directive is processed ahead of outlining function of PARALLEL directive. Therefore, the AST outlining function will consider these two variables are undeclared and report an error, although we know they will be added as parameters in the outlined function.