Fix the loop upper bound updating step
LLVM OpenMP runtime uses a variable to control the local loop upper bound for each thread and compares it with the original upper bound.
When it's an incremental loop, the changing upper bound should be always less than the actual upper bound.
When it's an decremental loop, it's should be always larger than the actual upper bound.