Add two kmpc ids to the outlined function
After generating the outlined function, the two kmpc parameters are added. This approach doesn't change anything in the outliner module.
This insertion must be performed before creating the forward declaration.

A separate insertion helper is created in omplowering module. It could be used to both append or prepend the parameter.

For now, the transformation of SINGLE directive always use a kmpc function call to get the global tid. It will be addressed soon.