Add a meaningful suffix to the outlined function name
The generated outlined function name is unique but not meaningful.
To resolve issue #66, the added suffix is "<enclosing function name>__<line number of the
original statement>__".

For example, the previous name "OUT__1__5599__" would be changed to "OUT__1__5599__main__10__". It indicates the original OpenMP code is in the main function at line 10.