Implement a post-processing module to handle the outlined functions
If the global variables are not static, they need to be set to extern without definition in the new file. Otherwise, the compiler would complain that they are redefined.

For example, in the original file, there are global variables:
int a = 1;
static b = 2;

In the new file, they would be:
extern a;
static b = 2;