Implement the initial version of omp target offloading using LLVM runtime
This commit is addressing issue #71.
Both outlined driver and CUDA kernel are stored in the same file.
Device id, number of threads per block, and number of blocks are set to a default value.
The size and type of mapping data are not handled yet.
The helper function to register .cubin file is stored in a separate file.