Update the transformation of omp target offloading
The offloading entries are stored in an global array. They will be registered by a generated helper. The entry key is a unique integer instead of function address.

Multiple GPU kernels can be handled properly. All the kernels are stored in one outlined function file. All the entries are set up in the main file.