Restore the CUDA language support
In commit https://github.com/passlab/rexompiler/commit/2073f14fe5fe4218f1491bc15e90d95a83e6263e,
we removed the pre-included CUDA headers.
However, they are required to support CUDA source code. It's necessary to copy the file "preinclude-cuda.h" and specify the path to CUDA headers in the system.

Since the CUDA 7.5 header package in REX was removed, to build REX compiler with CUDA support, we need to run:
./configure --prefix=$REX_ROOT/rex_install --with-boost=/usr --with-boost-libdir=$BOOST_LIB --enable-languages=c,c++,cuda,fortran --with-cuda-inc=/opt/cuda/include --disable-tests-directory --disable-tutorial-directory

"--with-cuda-inc=<path-to-CUDA-headers>" should be modified based on the actual CUDA toolkit installation.