Fix omp lowering
Last synchronization with ROSE brought a change in the file "src/frontend/SageIII/sageInterface/sageBuilder.C".

When creating a new source file for the outlined function, it was attaching some necessary preprocessing info to the new file.
In the latest commits from ROSE, during outlining, it calls a secondary pass to duplicate the original source file, which contains all the OpenMP source code, and then repeat the whole processing from scratch again. This secondary pass will parse the source code, create the AST and lower the AST.
The lowering will trigger the secondary pass again. Therefore, it falls to an infinite loop here.

As a temporary fix to issue #104, we reverted this particular change for now. More details should be investigated later.