Refactor UPIR Loop Parallel statement
A new UPIR node SgUpirLoopParallelStatement is added, which has worksharing, simd, and taskloop as its three children node to indicate how the loop should be parallelized.
It has a SgUpirLoopStatement as child node for the canonical loop.
The worksharing node SgUpirWorksharingStatement is refactored from the previous SgUpirLoopParallelStatement/SgOmpForStatement.

For now, all the clauses should be kept with worksharing, simd, and taskloop statements to reuse existing transformation with minor changes.