Fix transformation of reduction on GPU
We still have to split the reduction into two steps:
1. Reduce within blocks.
2. Reduce partial results from all blocks.

Because there's no easy way to synchronize on a grid level, we don't know when step 1 is completed.
Therefore, step 2 must happen in a parent kernel on GPU, or on the host, after the kernel in step 1 exits. Since the workload in step 2 is very light, it's not worth to launch another kernel, it's better to do it on the host.